Octave control working again
But I removed the key tags, because I thought they aren’t used. Which
is wrong. So I must add that again in the next commit.
